pyproject.toml
[tool.poetry]
name = "gen-bootstrap"
version = "0.1.0"
description = "A scaffold and CLI for generative AI projects using the Google ADK on GCP."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

packages = [
    { include = "cli" },
    { include = "adk" },
    { include = "config" },
    { include = "tools" },
    { include = "utils" },
    { include = "main.py" },
    { include = "test_client.py" },
    { include = "__init__.py" } # Include the root __init__.py
]

[tool.poetry.scripts]
"gen-bootstrap" = "cli.main:app" # Path updated for flat layout

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
google-cloud-aiplatform = "^1.38.0" # Or the latest compatible version
google-cloud-logging = "^3.8.0" # Or the latest compatible version
google-cloud-trace = "^1.10.0" # Or the latest compatible version
google-cloud-secret-manager = "^2.16.2" # Or the latest compatible version
typer = {extras = ["all"], version = "^0.15.2"} # Includes Click, Updated for fastmcp compatibility
fastapi = "^0.115.2" # Or the latest compatible version
uvicorn = {extras = ["standard"], version = "^0.34.0"} # Updated for google-adk compatibility
ttok = { git = "https://github.com/j3brns/token_count_trim.git" }
# Google ADK dependency will be added here once available/specified
gradio = "^5.25.2"
python-dotenv = "^1.1.0"
google-adk = "^0.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4" # Or the latest compatible version
pytest-mock = "^3.12.0" # Or the latest compatible version
pre-commit = "^3.6.0" # Or the latest compatible version
black = "^23.12.1" # Or the latest compatible version
isort = "^5.13.2" # Or the latest compatible version
flake8 = "^6.1.0" # Or the latest compatible version
nbstripout = "^0.6.1" # Or the latest compatible version
jupyterlab = "^4.0.11" # Or the latest compatible version

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
